requirements.txt
absl-py==0.11.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
bcrypt==3.2.0
bleach @ file:///tmp/build/80754af9/bleach_1611254400839/work
cachetools==4.2.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255207413/work
chardet==4.0.0
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.1.0
email-validator==1.1.2
entrypoints==0.3
flake8==3.8.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-gTTS==0.18
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gTTS==2.2.2
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1604101327119/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==1.1.0
jedi==0.17.0
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///C:/ci/jupyter_core_1606130908768/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib==3.3.3
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
notebook @ file:///C:/ci/notebook_1611348264852/work
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
panda==0.3.1
pandas==1.2.1
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.1.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.2.0
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
python-dateutil==2.8.1
pytz==2020.5
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.6.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///C:/ci/six_1605205426665/work
SQLAlchemy @ file:///C:/ci/sqlalchemy_1612467804654/work
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
torch==1.7.1+cpu
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.8.2+cpu
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
